requirements.txt
asgiref==3.2.10
certifi==2020.6.20
chardet==4.0.0
cloudinary==1.22.0
dj-database-url==0.5.0
dj-rest-auth==1.1.2
Django==3.1.2
django-admin-interface==0.18.3
django-colorfield==0.5.0
django-cors-headers==3.5.0
django-environ==0.4.5
django-filter==2.4.0
django-flat-responsive==2.0
django-flat-theme==1.1.4
django-heroku==0.3.1
django-phonenumber-field==5.0.0
django-rest-auth==0.9.5
djangorestframework==3.12.1
gunicorn==20.0.4
idna==2.10
importlib-metadata==2.0.0
Markdown==3.2.2
mysqlclient==2.0.1
phonenumberslite==8.12.11
Pillow==7.2.0
psycopg2==2.8.6
pytz==2020.1
requests==2.25.1
sentry-sdk==1.5.1
six==1.15.0
sqlparse==0.3.1
urllib3==1.25.11
whitenoise==5.2.0
zipp==3.3.0
